pyproject.toml
[tool.poetry]
name = "tensorflow-practice"
version = "0.1.0"
description = "TensorFlow Practice"
authors = ["GiDaeHyeon <spliter2157@gmail.com>"]
readme = "README.md"
packages = [{include = "tensorflow_practice"}]

[tool.poetry.dependencies]
python = "^3.9"
tensorflow-gpu = "2.6.1"
transformers = "^4.23.0"
PyYAML = "^6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
